+++ examples/1_simple.py
@@ -5,10 +5,10 @@ from pydantic_resolve import resolve
 class Student(BaseModel):  # <==== can resolve fields
     name: str
 
-    greet: str = ''
-    async def resolve_greet(self):
+    answer: str = ''
+    async def resolve_answer(self):
         await asyncio.sleep(1)  # mock i/o
-        return f'hello {self.name}'
+        return f'{self.name} say the answer is 42'
 
 async def main():
     students = Student(name='john' )