+++ examples/5_sqlalchemy_loaderdepend.py
@@ -11,7 +11,7 @@ from sqlalchemy.ext.asyncio import create_async_engine
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
-from pydantic_resolve import resolve
+from pydantic_resolve import Resolver, LoaderDepend
 from pprint import pprint
 
 engine = create_async_engine(
@@ -48,18 +48,28 @@ async def insert_objects() -> None:
             session.add_all(
                 [
                     Task(id=1, name="task-1"),
-                    Task(id=2, name="task-2"),
-                    Task(id=3, name="task-3"),
 
                     Comment(id=1, task_id=1, content="comment-1 for task 1"),
-                    Comment(id=2, task_id=1, content="comment-2 for task 1"),
-                    Comment(id=3, task_id=2, content="comment-1 for task 2"),
 
                     Feedback(id=1, comment_id=1, content="feedback-1 for comment-1"),
-                    Feedback(id=2, comment_id=1, content="feedback-1 for comment-1"),
-                    Feedback(id=3, comment_id=1, content="feedback-1 for comment-1"),
+                    Feedback(id=2, comment_id=1, content="feedback-2 for comment-1"),
+                    Feedback(id=3, comment_id=1, content="feedback-3 for comment-1"),
                 ]
             )
+
+async def insert_new_objects() -> None:
+    async with async_session() as session:
+        async with session.begin():
+            task_1 = (await session.execute(select(Task).filter_by(id=1))).scalar_one()
+            task_1.name = 'task-1 x'
+            session.add(task_1)
+            session.add_all(
+                [
+                    Comment(id=2, task_id=1, content="comment-2 for task 1"),
+                    Feedback(id=4, comment_id=2, content="test"),
+                ]
+            )
+
 
 # =========================== Pydantic Schema layer =========================
 class FeedbackLoader(DataLoader):
@@ -72,7 +82,6 @@ class FeedbackLoader(DataLoader):
                 dct[row.comment_id].append(FeedbackSchema.from_orm(row))
             return [dct.get(k, []) for k in comment_ids]
 
-
 class CommentLoader(DataLoader):
     async def batch_load_fn(self, task_ids):
         async with async_session() as session:
@@ -84,8 +93,6 @@ class CommentLoader(DataLoader):
                 dct[row.task_id].append(CommentSchema.from_orm(row))
             return [dct.get(k, []) for k in task_ids]
 
-feedback_loader = FeedbackLoader()
-comment_loader = CommentLoader()
 
 class FeedbackSchema(BaseModel):
     id: int
@@ -101,7 +108,7 @@ class CommentSchema(BaseModel):
     content: str
     feedbacks: Tuple[FeedbackSchema, ...]  = tuple()
 
-    def resolve_feedbacks(self):
+    def resolve_feedbacks(self, feedback_loader = LoaderDepend(FeedbackLoader)):
         return feedback_loader.load(self.id)
 
     class Config:
@@ -112,52 +119,43 @@ class TaskSchema(BaseModel):
     name: str
     comments: Tuple[CommentSchema, ...]  = tuple()
     
-    def resolve_comments(self):
+    def resolve_comments(self, comment_loader = LoaderDepend(CommentLoader)):
         return comment_loader.load(self.id)
 
     class Config:
         orm_mode = True
 
-async def main():
+async def init():
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
 
-    await insert_objects()
+async def query_tasks():
+    print('---------------- query --------------')
     async with async_session() as session:
         tasks = (await session.execute(select(Task))).scalars().all()
         task_objs = [TaskSchema.from_orm(t) for t in tasks]
-        resolved_results = await resolve(task_objs)
-        to_dict_arr = [r.dict() for r in resolved_results]
-        pprint(to_dict_arr)
+        resolved_results = await Resolver().resolve(task_objs)
+        arr = [r.dict() for r in resolved_results]
+        pprint(arr)
+
+async def main():
+    """
+    almost the same with previous demo except using LoaderDepend to isolate the cache by async contextvar.
+    result of first and second shall be different.
+    """
+    await init()
+    await insert_objects()
+    # first
+    await query_tasks()
+
+    # check the update and insert
+    await insert_new_objects()  
+
+    # second
+    await query_tasks()
     await engine.dispose()
+
 
 loop = asyncio.get_event_loop()
 loop.run_until_complete(main())
 loop.close()
-
-# expected
-# [{'comments': [{'content': 'comment-1 for task 1',
-#                 'feedbacks': [{'comment_id': 1,
-#                                'content': 'feedback-1 for comment-1',
-#                                'id': 1},
-#                               {'comment_id': 1,
-#                                'content': 'feedback-1 for comment-1',
-#                                'id': 2},
-#                               {'comment_id': 1,
-#                                'content': 'feedback-1 for comment-1',
-#                                'id': 3}],
-#                 'id': 1,
-#                 'task_id': 1},
-#                {'content': 'comment-2 for task 1',
-#                 'feedbacks': [],
-#                 'id': 2,
-#                 'task_id': 1}],
-#   'id': 1,
-#   'name': 'task-1'},
-#  {'comments': [{'content': 'comment-1 for task 2',
-#                 'feedbacks': [],
-#                 'id': 3,
-#                 'task_id': 2}],
-#   'id': 2,
-#   'name': 'task-2'},
-#  {'comments': [], 'id': 3, 'name': 'task-3'}]